pyproject.toml
[tool.poetry]
packages = [{ include = "video_splitter" }]

name = "video-splitter"
version = "0.1.0"
description = ""
authors = ["Isaac Pedisich <isaac@pipe.com>"]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.10.0"
PyYAML = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
split = "video_splitter.splitter:main"
